Add nginx configuration
This nginx configuration talks acts as a reverse proxy to a uWSGI
server. Incoming traffic on port 80 (because we intend to terminate TLS
before this point), and traffic is forwarded to uWSGI on port 3031.

This
configuration is somewhat specialized for docker-compose since the uWSGI
container's hostname matches its service name (i.e., uwsgi).

For Kubernetes the hostname would be localhost since all containers in a
pod share the same network namespace.

This configuration is installed by the Dockerfile into the container. It
tells nginx to setup a TCP-based socket to talk to a uWSGI server (that
lives in another container).